+++ src/Core/src/Traits/Config/AliasTrait.php
@@ -11,6 +11,8 @@ declare(strict_types=1);
 
 namespace Spiral\Core\Traits\Config;
 
+use Spiral\Core\Exception\Container\ContainerException;
+
 /**
  * Provides aliasing ability for config entities.
  *
@@ -20,7 +22,13 @@ trait AliasTrait
 {
     public function resolveAlias(string $alias): string
     {
+        $antiCircleReference = [];
         while (is_string($alias) && isset($this->config) && isset($this->config['aliases'][$alias])) {
+            if (\in_array($alias, $antiCircleReference, true)) {
+                throw new ContainerException("Circle reference detected for alias `$alias`.");
+            }
+            $antiCircleReference[] = $alias;
+
             $alias = $this->config['aliases'][$alias];
         }