+++ src/Scaffolder/src/Declaration/MigrationDeclaration.php
@@ -18,6 +18,7 @@ use Spiral\Reactor\DependedInterface;
 
 /**
  * Migration declaration
+ * @deprecated since v2.10. Will be moved to spiral/cycle-bridge and removed in v3.0
  */
 class MigrationDeclaration extends ClassDeclaration implements DependedInterface
 {